feat: factorisation du code dans simulation
